chore: move webbtc and specify scope
