+++ src/extension/content-script/onend.js
@@ -37,7 +37,7 @@ async function init() {
     }
   });
 
-  // message listener to listen to inpage webln calls
+  // message listener to listen to inpage webln/webbtc calls
   // those calls get passed on to the background script
   // (the inpage script can not do that directly, but only the inpage script can make webln available to the page)
   window.addEventListener("message", (ev) => {
