+++ src/extension/inpage-script/webln.js
@@ -1,5 +1,5 @@
+import WebBTCProvider from "../ln/webbtc";
 import WebLNProvider from "../ln/webln";
-import WebBTCProvider from "../webbtc";
 
 if (document) {
   window.webln = new WebLNProvider();
